refactoring "set person away"
